Reorganize SendToTVCoordinator
