+++ Packages/App/Sources/AppUIMain/Views/Profile/iOS/SendToTVQRScanView+iOS.swift
@@ -28,6 +28,7 @@
 import CommonUtils
 import SwiftUI
 
+// FIXME: ###, iOS QR instructions
 struct SendToTVQRScanView: View {
     let onDetect: (String) -> Void
