+++ Packages/App/Sources/CommonUtils/Views/PasscodeInputView.swift
@@ -25,7 +25,6 @@
 
 import SwiftUI
 
-// FIXME: ###, layout + instructions
 public struct PasscodeInputView: View {
     private let length: Int
