+++ Packages/App/Sources/CommonUtils/Views/QRScanView.swift
@@ -30,7 +30,6 @@ import SwiftUI
 import UIKit
 import Vision
 
-// FIXME: ###, instructions
 public struct QRScanView: UIViewControllerRepresentable {
     private let onDetect: (String) -> Void
